hierarchy module now just deserializes datacleaning.py artifacts
