+++ docs/icdcodex.data.rst
@@ -1,0 +1,10 @@
+icdcodex.data package
+=====================
+
+Module contents
+---------------
+
+.. automodule:: icdcodex.data
+    :members:
+    :undoc-members:
+    :show-inheritance:
