+++ icdcodex/data/__init__.py
@@ -1,0 +1,1 @@
+"""submodule for holding data artifacts, including ICD9/10 hierarchys in serialized networkX format and precomputed embeddings"""
